+++ pytest_django/plugin.py
@@ -5,6 +5,7 @@ test database and provides some useful text fixtures.
 """
 
 import contextlib
+from distutils.util import strtobool
 import inspect
 from functools import reduce
 import os
@@ -67,6 +68,9 @@ def pytest_addoption(parser):
     group._addoption('--migrations',
                      action='store_false', dest='nomigrations', default=False,
                      help='Enable Django 1.7+ migrations on test setup')
+    group._addoption('--django-debug',
+                     action='store', dest='djangodebug', default='False',
+                     help='Configure the DEBUG setting. Defaults to False')
     parser.addini(CONFIGURATION_ENV,
                   'django-configurations class to use by pytest-django.')
     group._addoption('--liveserver', default=None,
@@ -336,7 +340,8 @@ def django_test_environment(request):
         from django.conf import settings as dj_settings
         from django.test.utils import (setup_test_environment,
                                        teardown_test_environment)
-        dj_settings.DEBUG = False
+        if request.config.getvalue('djangodebug') != 'None':
+            dj_settings.DEBUG = bool(strtobool(request.config.getvalue('djangodebug')))
         setup_test_environment()
         request.addfinalizer(teardown_test_environment)